Rename Accordion; fix typo
